// src/types/Facility.ts
export interface Facility {
  ccn: string;
  providerName: string;
  providerAddress: string;
  city: string;
  state: string;
  zipCode: string;
  certifiedBeds: string;
  averageResidentsPerDay: string;
  overallRating: string;
  healthInspectionRating: string;
  staffingRating: string;
  qmRating: string;
  location: string;
}